requirements.txt
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
datasets==3.3.2
evaluate==0.4.3
accelerate==1.4.0
sentencepiece==0.2.1
jiwer==3.0.5
Pillow==10.4.0
numpy==2.1.1
PyYAML==6.0.2
tqdm==4.67.1
scikit-learn==1.6.1
